+++ sdk/core/azure-core/CHANGELOG.md
@@ -5,6 +5,7 @@
 ### Features Added
 
 - Added new error type `IncompleteReadError` which is raised if peer closes connection without sending complete message body.
+- add kwargs to the methods for `iter_raw` and `iter_bytes`  #21529
 
 ### Breaking Changes